Move all nanobind code to nanobind.cc
