+++ cpp/nanobind/python_methods.h
@@ -7,7 +7,6 @@
 #ifndef XGRAMMAR_NANOBIND_PYTHON_METHODS_H_
 #define XGRAMMAR_NANOBIND_PYTHON_METHODS_H_
 
-#include <nanobind/nanobind.h>
 #include <xgrammar/xgrammar.h>
 
 #include <optional>
@@ -27,8 +26,6 @@ TokenizerInfo TokenizerInfo_Init(
 );
 
 std::string TokenizerInfo_GetVocabType(const TokenizerInfo& tokenizer);
-
-std::vector<nanobind::bytes> TokenizerInfo_GetDecodedVocab(const TokenizerInfo& tokenizer);
 
 bool GrammarMatcher_FillNextTokenBitmask(
     GrammarMatcher& matcher,
